Refactor the use of reflection macros
The reflection layer now uses a set of header files
that avoid excessive includes and macro processing.
This is a preliminary work and must be checked.

The test programs have been updated.

The example programs will be updated later.

This migration has only be done for datatools.
The rest of Bayeux (geomtools, genbb_help) is broken.

The next steps will be:
- fix other bayeux modules using the same strategy
- remove alias macros (camp wrapping) and privatize
  more strictly registration macros