Fix macro with empty arg(s)
Usage of macro like:

MY_MACRO( foo, /*bar*/ )

where 'bar' is not supposed to be used is not correct in C++98.
The intention was to tag the bar parameter as unused
from the signature of a function/method.

One now uses the explicit "__attribute__((unused))" tag:

MY_MACRO( foo, bar __attribute__((unused)) )

that is recognized by GCC and Clang and should not introduced
any side effect. This technique must be reviewed (Ben?).