Use the angular_range class for various shapes
A   few   1D   and   2D   shapes  now   uses   attribute(s)   of   the
geomtools::angular_range class  to implement  partial angle  (polar or
azimuthal).  This technique will be  propagated to other shape classes
in the future.  This is a long  term solution to a bug found by Arnaud
in some shape classes and related to partial angular domain.