Use GSL Imported targets
CMake provides imported targets for GSL in its builtin FindGSL module. 
As we require CMake 3.3 and this provides the needed functionality,
migrate to using the module, allowing upstream removal of the GSL resource
in Cadfael.

Additionally update re-finding of GSL in BayeuxConfig to use the module's
functionality to recreate the imported targets.