Implement serialization of display_data objects from generated vertexes
It is  now possible  to store,  in a  Boost/archive formatted  file, a
geomtools::display_data  object  that  contains  the  rendering  of  a
collection of vertexes generated from the genvtx_inspector.

The file can  then be loaded from the geomtools  inspector in order to
benefit of  full navigation features and at the same time be able to
superimpose the rendering of vertexes on a virtual geometry setup.

This make the validation of vertex generation algorithm far more easy
and convenient, particularly in very 'dense' geometries for which basic
display features in the genvtx inspector are not very practical/confortable.

Readme in example ex01 in updated to explain how to use this new feature.