Refactor I/O save/load utilities for '(multi_)properties' classes
Both 'datatools::properties' and 'datatools::multi_properties' classes
use a dedicated nested class ::config with read/write functionalities
for I/O on configuration files.

The technique is the same for both classes and 'datatools::multi_properties::config'
class makes use of the 'datatools::properties::config' class.
Old interface with the activation of boolean flags has been removed and replaced
by constructors with specific enumerated flags responsible of the activation
of parsing and/or writing functionalities.

Additional code has neen introduced in the parsing/read methods in order to
obtain better reports on syntax error.

The system is far much easier to maintain in this way.
The format of configuration files remains the same of course.