Fix a subtle bug in variant support
The handling of a certain kind or error in the syntax of configuration
files with variant supports was broken and make the library choose the
default parameter value in place of reporting an error. This has been fixed.

Refactor and clean several pieces of code.

Make the kernel initialization more flexible.