Add bxg4_seeds application.
The bxg4_seeds application generates sets of independant seeds which are
usable by different runs of the bxg4_production program. It is ensured
that each run has its own set of seeds which is not reused by any other
run in order to guarantee the statistical independance of the simulation
output.

Suppose the user plans to launch a large set of simulation runs (sequentially
or in parallel using a computing cluster).
He/she first generate the seeds for the set of scheduled runs using:

  $ bxg4_seeds -l notice -n 1000 -s 0 -d /tmp/simulation -p "run_%n/seeds.def"

The "run_1/seeds.def" to "run_999/seeds.def" file contain their own set of 4 seeds
which will be used by the bxg4_production program:

  $ bxg4_production --input-prng-seeds-file="run_0/seeds.def" ... --output-data-file="run0.data.gz"
  $ bxg4_production --input-prng-seeds-file="run_1/seeds.def" ... --output-data-file="run1.data.gz"
  $ bxg4_production --input-prng-seeds-file="run_2/seeds.def" ... --output-data-file="run2.data.gz"
  ...
  $ bxg4_production --input-prng-seeds-file="run_999/seeds.def" ... --output-data-file="run999.data.gz"

All runs are thus made statistically independant.

For convenience, a file is produced which contains the list of run numbers and associated seeding file:

  0 /tmp/toto/run_0/simulation/config/seeds.def
  1 /tmp/toto/run_1/simulation/config/seeds.def
  ...
  999 /tmp/toto/run_999/simulation/config/seeds.def

Each run seeding file used the following format:

  {EG=994968840; MGR=71833851; SHPF=405254415; VG=604639512}

where:

 - "EG" stands for "event generator"
 - "VG" stands for "vertex generator"
 - "MGR" stands for "Geant4 simulation manager"
 - "SHPF" stands for "Step hit processor factory" (pre-digitization)