Add new placement strategy for internal items
The "geomtools::model_with_internal_items_tools" class,  which is used
to position daughter/child volumes inside a mother logical volume, has
new features.

First of all, it is still  possible to use the 'legacy' compact syntax
to define the placement of a daughter volume within its mother volume:

Example:
  internal_item.labels : string[1] = "Child"
  internal_item.model.Child     : string = "my_geometry.model"
  internal_item.placement.Child : string = "0.0 10.0 -13.0 cm / z 45 (degree)"

The new "geomtools::placement::builder"  class is used to  support a new
syntax and a new placement algorithm.

With  the  new syntax,  each  coordinate  (x, y  or  z)  of the  child
placement can be set individually.  Example:

  internal_item.placement.Child.x : real as length = 0.0 cm
  internal_item.placement.Child.y : real as length = 10.0 cm
  internal_item.placement.Child.z : real as length = -13.0 cm

as well as rotation parameters:

  internal_item.placement.Child.rotation_axis  : string = "z"
  internal_item.placement.Child.rotation_angle : real as angle = 45 degree

This make easy to add variant support to such parameters, like in:

  internal_item.placement.Child.rotation_angle : real as angle = @variant(geometry:angle|0 degree)

The  'builder' is  also able  to psotion  a daughter  volume inside  a
mother  logical volume  by using  a  relative placement  of the  child
volume with respect  to the limits/borders of the  mother volume. This
is named the 'Child/mother gap' positioning, with 6 strategy supported
for each of the x, y, z axis.

The "Child/mother gap mode" supports the following strategies:

 * "center_to_min" : the distance between the origin of the daughter volume
   and the minimum limit of the mother volume (generally min x/y/z) is fixed
 * "center_to_max" : the distance between the origin of the daughter volume
   and the maximum limit of the mother volume (generally max x/y/z) is fixed
 * "min_to_min" : the distance between the minimum limit of the of the daughter volume
   and the minimum limit of the mother volume (generally min x/y/z) is fixed
 * "max_to_min" : the distance between the maximum limit of the of the daughter volume
   and the minimum limit of the mother volume (generally min x/y/z) is fixed
 * "min_to_max" : the distance between the minimum limit of the of the daughter volume
   and the maximum limit of the mother volume (generally max x/y/z) is fixed
 * "max_to_max" : the distance between the maximum limit of the of the daughter volume
   and the maximum limit of the mother volume (generally max x/y/z) is fixed

Example: the daughter volume is positionned on the floor (along the Z direction) of its mother volume

  internal_item.placement.Child.x : real as length =  10.0 cm # explicit X position
  internal_item.placement.Child.y : real as length = -20.0 cm # explicit Y position
  internal_item.placement.Child.z.gap_mode     : string         = "min_to_min"
  internal_item.placement.Child.z.gap_distance : real as length = 0.0 cm # Gravity

             ^ z
  mother     :
  +----------------------+
  |                      |
  |                      |
  |                      |
  |                      |
  |       +------+       |
  |       |child |       |
  +-------+------+-------+

Note  that the  use of  the new  "Child/mother gap  mode" can  lead to
undesirable  behaviour  (overlap) if  combined  with  rotation of  the
daughter volume inside its mother. It  is expected that this mode will
not  be used  in coincidence  with explicit  rotation of  the daughter
volumes. This is because the gap distance between borders of child and
mother volumes  is, for a  given axis, computed from  their respective
"stackable data"  or "bounding  data".  Rotating the  daughter volumes
will almost  always break the  calculation of the gap  mode placement,
depending of the actual shape of the daughter volume.

File 'doc/images/placement_modes.fig' defines the placement parameters
depending of the gap mode.

Example "example/ex00" is modified to illustrate the use of this new features.