Many minor/major mods
Due to recent svn repository problems, it was not possible
to commit step-by-step changes for about 3 weeks.

The consequence is this big set of nested mods:
- some are only kozmetics
- some are minor mods
- there are a few changes in some interface of I/O methods (ioutils.h)
  and associated mods in some other files.
- there are new classes and utilities in the "include/datatools/ui/"
  and "include/datatools/configuration" sectors, particularly because
  we need new generic fonctionalities in Vire.
- there are some bugfixes
- there are new test programs

Currently, my Vire and Falaise trunks seems ok with this release
but we need to check if no side effect is introduced.
Hope to fix any issue within a few days.

Apologizes for this big nasty summer commit!

frc