Add dynamic mode in service manager
Formerly, it was not possible to load a new service
in a service manager object when it was already initialized.
It is now possible to configure the service manager in such a way
a service can be dynamically added even after initialization.
The standard behaviour is unchanged (dynamic loading is forbidden)
but a special flag at construction is possible
to allow dynamic loading and dropping.
In principle, this change should have no impact on existing code.

The method "service_manager::build_list_of_services" is also provided
to build the list of names of the services currently embedded in the manager.
Limited filtering is provided.