Add a utility script (preliminary).
The `bxextract_table_of_objects` application aims to extract a list
of named objects with associated metadata from a configuration
file using the `datatools::multi_properties` format.

The generated list can be output to ReSt or CSV format
for external clients (documentation support...).

Note: Could be better implemented in C++ in the future.