Refactor dependency support of the variant system
The  first implementation  of  variants had  a malformed  "long-range"
dependency  support. "Long  range"  dependency  support concerns  some
possible correlations between  variant configuration parameters hosted
in different  branches of  the variant  repository or  registries.

For  example,  Changing  the  layout  of  the  geometry  through  some
parameters in a  "Geometry" variant registry may  have consequences on
the  variant   parameters  and  usecases  published   from  a  "Vertex
generators" registry.   However, when the "Geometry"  variant registry
is defined,  it is  not possible  to know  in advance  what additional
services  will possibly  depends on  its parameters.   When a  "Vertex
generators" registry  is integrated  on top  of the  former "Geometry"
variant  registry, it  is now  possible to  define a  dependency logic
associated to some of the new variants.

Example:

The  vertex  generator  associated  to a  calibration  source  can  be
instantiated and activated if and  only if the geometry describes such
calibration source in  the setup. The choice of  this vertex generator
is thus automatically published to  the user when he/she activates the
proper variant from the geometry.

A better  model of such  dependency is  now introduced. It  enables to
express complex dependency logics between variant objects (parameters,
parameter values...).

The  syntax to  setup such  kind  of dependencies  is modified.   This
implies some  changes in some  Falaise config files for  example. This
will be done as soon as possible.