Addons and refactoring
1 - Utilities for URN manipulation

URN is a standard way of addressing objects (or concepts)
thanks to string identifier. Bayeux/datatools provides some new classes
to parse, store and manipulate URNs.
It is also possible to associate some basic informations
to a given URN and store them in some dictionaries that can be
traversed by client to pickup useful informations such as:
- object with URN="urn:xxx:yyy:4.3" depends on objects
  with URN="urn:aaa:bbb:1.0" and URN="urn:aaa:ccc:2.4"
- configuration with URN="urn:mynemo:daq:config:1.0" is associated
  to filesystem path "${MYNEMO_CONFIG_DIR}/daq/1.0-test/run.conf"

This provides a technique to represent formally some special relationship between
various objects/instances :
- dependencies
- filesystem path resolution

This enables to hide some implementation details, concentrating on the logics
on which object should be used within a given software.

New classes:
- datatools::urn
- datatools::urn_info
- datatools::urn_to_path_resolver_service
- datatools::urn_db_service
- datatools::urn_query_service

2 - Datatools' kernel singleton:

The kernel is refactored with an embedded service_manager.
This manager is the host for various 'system' services which
should be accessible from anywhere in the code and for clients
(apps, plugins...) because the kernel is a static singleton (as before).

- the former kernel's 'library_info' object is now wrapped in
  a 'library_query_service' (new class)
- an 'urn_query_service' instance is used to register various
  informations about URNs associated to various objects of software
  components addressed through some blessed URN scheme loaded from
  tables.
  - Instances of the 'urn_to_path_resolver_service' new class are
    typically used to find the filesystem path of a file resource
    addressed by its URN. They can freely register themselves to
    the kernel, making their lookup tables available with global scope.
  - Instances of the 'urn_db_service' new class are used to describe
    the dependencies between objects address through their URNs. They
    can freely register themselves to
    the kernel, making their lookup tables available with global scope.

3 - Refactor some parts of the datatools library and kernel initialization and
    shutdown.

4 - Various fixes.

This is a preliminary release that probably breaks Bayeux.
Bayeux needs to be fixed to adapt to these changes. It is on the way.