Relocate the path resolution algo in libinfo
Originaly, the path resolution algo was embedded as a datatools' kernel
method, on top of its embedded library_info instance. It is now moved
as a library_info class's method together with other utilities.
This simplifies the implementation for maintenance and reusability.
The path resolving functionnalities are thus usable, is needed, directly from
a library_info instance (and thus a library_query service) without
the help of the static kernel.