Experimental fix for XML read errors
For Boost 1.66 and newer, tests reading from XML Boost.Serialization
archives throw exceptions when the datatools::data_reader (and hence
io_factory) is destructed. This is traced to the destructor of the held
instance of boost::archive::xml_iarchive, which throws an
input_stream_error exception.

Based on information from:

- boostorg/serialization#82
- boostorg/serialization#99
- boostorg/serialization#109

writing produces the correct XML, but reading fails. Using the info
from boostorg/serialization#99, the held stream is explicitly moved
to EOS before destructing. This fixes all bar two errors in tests
for Boost 1.68 which appear unrelated to the XML issue.